third_party/blink/web_tests/platform/generic/http/tests/devtools/console/console-on-animation-worklet-expected.txt

CONSOLE DEBUG: line 1: Debug
CONSOLE INFO: line 2: Information
CONSOLE MESSAGE: line 3: Log
CONSOLE WARNING: line 4: Warning
Tests console output from AnimationWorklet.

Message count: 4
console-worklet-script.js:1 Debug
console-worklet-script.js:2 Information
console-worklet-script.js:3 Log
console-worklet-script.js:4 Warning
(anonymous) @ console-worklet-script.js:4
